+++ packages/functions-testing.experimental/src/index.ts
@@ -31,3 +31,5 @@ export {
   type StubBuilderFor,
   type StubClient,
 } from "./mock/createMockClient.js";
+
+export { createMockObjectSet } from "./mock/createMockObjectSet.js";