+++ packages/react-components/src/action-form/fields/FormFieldRenderer.tsx
@@ -20,6 +20,8 @@ import type { RendererFieldDefinition } from "../FormFieldApi.js";
 import { CustomField } from "./CustomField.js";
 import { DatetimePickerField } from "./DatetimePickerField.js";
 import { DropdownField } from "./DropdownField.js";
+import { FilePickerField } from "./FilePickerField.js";
+import { NumberInputField } from "./NumberInputField.js";
 import { RadioButtonsField } from "./RadioButtonsField.js";
 import { TextAreaField } from "./TextAreaField.js";
 import { TextInputField } from "./TextInputField.js";
@@ -118,13 +120,46 @@ function renderFieldComponent(
         />
       );
     case "NUMBER_INPUT":
+      // TODO: Use coerceFieldValue
+      return (
+        <NumberInputField
+          id={fieldDefinition.fieldKey}
+          value={typeof value === "number" ? value : null}
+          onChange={onChange}
+          placeholder={fieldDefinition.placeholder}
+          {...fieldDefinition.fieldComponentProps}
+        />
+      );
     case "FILE_PICKER":
+      return (
+        <FilePickerField
+          id={fieldDefinition.fieldKey}
+          value={coerceToFileValue(value)}
+          onChange={onChange}
+          {...fieldDefinition.fieldComponentProps}
+        />
+      );
     case "OBJECT_SET":
       return <div>Unsupported field type: {fieldDefinition.fieldComponent}
       </div>;
     default:
       return assertUnreachableFieldComponent(fieldDefinition);
   }
+}
+
+// TODO: Move and share with `coerceFieldValue`
+function isFileArray(value: unknown[]): value is File[] {
+  return value.every((v) => v instanceof File);
+}
+
+function coerceToFileValue(value: unknown): File | File[] | null {
+  if (value instanceof File) {
+    return value;
+  }
+  if (Array.isArray(value) && isFileArray(value)) {
+    return value;
+  }
+  return null;
 }
 
 function assertUnreachableFieldComponent(value: never): never {