+++ packages/cli.common/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @osdk/cli.common
 
+## 0.32.0
+
+### Minor Changes
+
+- 322c5bc: Simulated release
+- 35f2f1a: Add Media inputs/outputs for Queries
+
+## 0.32.0-beta.5
+
 ## 0.32.0-beta.4
 
 ## 0.32.0-beta.3