+++ packages/client.test.ontology/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @osdk/client.test.ontology
 
+## 2.8.0
+
+### Minor Changes
+
+- 322c5bc: Simulated release
+- 35f2f1a: Add Media inputs/outputs for Queries
+
+### Patch Changes
+
+- Updated dependencies [322c5bc]
+- Updated dependencies [4bb9769]
+- Updated dependencies [f294f5a]
+- Updated dependencies [0df859a]
+- Updated dependencies [d284bf2]
+- Updated dependencies [4ef6adc]
+- Updated dependencies [60b5ffb]
+- Updated dependencies [8c30908]
+- Updated dependencies [4a856cb]
+- Updated dependencies [b68cebd]
+- Updated dependencies [35f2f1a]
+- Updated dependencies [61e33ab]
+  - @osdk/api@2.8.0
+
+## 2.8.0-beta.32
+
+### Patch Changes
+
+- @osdk/api@2.8.0-beta.32
+
+## 2.8.0-beta.31
+
+### Patch Changes
+
+- @osdk/api@2.8.0-beta.31
+
 ## 2.8.0-beta.30
 
 ### Patch Changes