+++ packages/create-widget/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @osdk/create-widget
 
+## 3.5.0
+
+### Minor Changes
+
+- 322c5bc: Simulated release
+- 35f2f1a: Add Media inputs/outputs for Queries
+
+## 3.5.0-beta.32
+
+## 3.5.0-beta.31
+
 ## 3.5.0-beta.30
 
 ## 3.5.0-beta.29