+++ packages/example-generator/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @osdk/example-generator
 
+## 0.16.0
+
+### Minor Changes
+
+- 322c5bc: Simulated release
+- 35f2f1a: Add Media inputs/outputs for Queries
+
+### Patch Changes
+
+- Updated dependencies [322c5bc]
+- Updated dependencies [322c5bc]
+- Updated dependencies [2d6652c]
+- Updated dependencies [35f2f1a]
+  - @osdk/create-app@2.8.0
+  - @osdk/create-widget@3.5.0
+
 ## 0.16.0-beta.2
 
 ### Minor Changes